Poker Hand Evaluator
This was originally designed to be used as part of a multiplayer poker
game in conjunction with a database, using LINQ and Web Services to
access it.
I have extracted the core hand evaluation functionality for display and
removed the database components.
This system was developed with ASP.NET for the UI.